Update occupation array in-place
Instead of rebuilding the whole occupation array, we now check if a change in bounds triggered a fixation of the variable. If so, we just update the occupation array and only restart the constraint propagation loop, instead of restarting from the very beginning.
Furthermore, we allow tightening the boundaries of a variable multiple times in a row when we were able to do a change. This way we can stay on one variable longer instead of discovering additional tightening opportunities after rerunning the whole algorithm.

Cycles: 145416006080
Instructions: 538279426450

Shaved off quite a lot of cycles and instructions with this one :^)