Update occupation array in-place in sum-constraint
Instead of a full restart we try to update the occupation array in-place so we can skip fully rebuilding it and jump straight to the alldifferent constraint check. This is the same optimization implemented in the alldifferent constraint check itself.
We now only fully restart for corner/symmetry constraints.

Cycles: 118185261745
Instructions: 356940050565

This change was more impactful than expected with an uplift of ~14%.