Replace most restart jumps with a do-while-loop + flag
Instead of immediately jumping back to the restart(_propagate_alldifferent) label we remember the kind of change in a flag and loop back to the start using a regular do-while-loop. The flag uses two bits: The first (LSB) marks the former restart_propagate_alldifferent and the second bit marks a full restart where we need to rebuild the occupation array. By utilizing bit-wise OR we prevent a later update to the flag to clobber already set bits.
Only the alldifferent constraint check still uses a goto-setup, we will try to change this next.

Cycles:  56917950209
Iterations: 148653302730

What is going on? This is way too overpowered. (>50% uplift?!)