Refactor repeating alldifferent improvements into loop
Instead of building an implicit loop by jumping back with a goto when trying to repeatedly improve a specific variable, we now feature a while-loop. It automatically stops at the end with a break, unless we call continue, leaving us with the same behavior as before. It also happens to be a bit faster.

Cycles: 56370145311
Iterations: 147896505813