Heuristicly select next variable to set while labeling
The labeling function now selects the next variable to try to set by using a heuristic function. After some trial and error we found, that the maximal lower bound is the best heuristic by far.
Which variables are left for selecting is tracked in an array, where items are removed using an unstable take in O(1) time.

Cycles: 3364493696
Instructions: 9236100378

With this change we have a speed up of 15x! This is actually hard to believe.