Try to bisect the variable range while labeling
Instead of trying out every possible value in a variables range when labeling, we now split the range by calling labeling twice, in case the range is greater than 3. This is supposed to detect bad branches early and reduce the overall number of explored leaves.

Cycles: 3158364801
Instructions: 9344737890

A modest improvement of ~4.5%.