Try another approach for iterative labeling
Instead of calling labeling recursively when `solve` succeeds, we push the result to a FIFO array. Every time we take a new block of vars and remaining values from the fifo and run `labeling` on it.
This version is only a proof of concept and very dub, as it does not memory reallocation or recycling. We just allocate 1GiB on the start and fill it up. This happens to be enough to run the benchmark `./magichex 4 3 14 33 30 34 39 6 24 20`.

However, as it turns out even this version which does not deal with any of the memory management complexities yet is still slower than our currently best recursive one. Both compiled with `DNDEBUG` compared (recursive left, iterative right)

Cycles: 2747223896 vs 2917492136
Instructions: 7472435221 vs 7485497817
Leaves: 384599 vs 383057

As we do not reorder the remaining vars array when putting back the index before returning, we explore a different number of leaves.

Sad that this did not work, it was quite a bit of work. ._.